refactor: redo how we do some processes between different OSes
