+++ src/app/data_harvester/processes/unix.rs
@@ -1,33 +1,37 @@
 //! Unix-specific parts of process collection.
 
-use hashbrown::HashMap;
+mod user_table;
+use cfg_if::cfg_if;
+pub use user_table::*;
 
-use crate::utils::error;
+cfg_if! {
+    if #[cfg(all(target_family = "unix", not(target_os = "linux")))] {
+        mod process_ext;
+        pub(crate) use process_ext::*;
 
-#[derive(Debug, Default)]
-pub struct UserTable {
-    pub uid_user_mapping: HashMap<libc::uid_t, String>,
-}
+        use sysinfo::System;
+        use super::ProcessHarvest;
 
-impl UserTable {
-    pub fn get_uid_to_username_mapping(&mut self, uid: libc::uid_t) -> error::Result<String> {
-        if let Some(user) = self.uid_user_mapping.get(&uid) {
-            Ok(user.clone())
-        } else {
-            // SAFETY: getpwuid returns a null pointer if no passwd entry is found for the uid
-            let passwd = unsafe { libc::getpwuid(uid) };
+        use crate::utils::error;
+        use crate::app::data_harvester::processes::*;
 
-            if passwd.is_null() {
-                Err(error::BottomError::QueryError("Missing passwd".into()))
-            } else {
-                // SAFETY: We return early if passwd is null.
-                let username = unsafe { std::ffi::CStr::from_ptr((*passwd).pw_name) }
-                    .to_str()?
-                    .to_string();
-                self.uid_user_mapping.insert(uid, username.clone());
+        pub fn get_process_data(
+            sys: &System, use_current_cpu_total: bool, unnormalized_cpu: bool, total_memory: u64,
+            user_table: &mut UserTable,
+        ) -> error::Result<Vec<ProcessHarvest>> {
+            cfg_if! {
+                if #[cfg(target_os = "macos")] {
+                    MacOSProcessExt::get_process_data(sys, use_current_cpu_total, unnormalized_cpu, total_memory, user_table)
+                } else if #[cfg(target_os = "freebsd")] {
+                    FreeBSDProcessExt::get_process_data(sys, use_current_cpu_total, unnormalized_cpu, total_memory, user_table)
+                } else {
+                    struct GenericProcessExt;
+                    impl UnixProcessExt for GenericProcessExt {}
 
-                Ok(username)
+                    GenericProcessExt::get_process_data(sys, use_current_cpu_total, unnormalized_cpu, total_memory, user_table)
+                }
             }
         }
+
     }
 }
